It works!
Rewrite a more sane version of extract_features. Make
fast_tree_to_cxx_score_iterate work. It currently includes a test rig to
make sure a generated corner detector is the same as the libcvd
implementation. Naturally, this only works if fast_N_features --N 9 is
used as the input.